package rename for all modules
