--- libModule1/src/main/java/com/infinum/libname/MyClass.java
@@ -1,4 +1,0 @@
-package com.infinum.libname;
-
-public class MyClass {
-}
