+++ libModule1/src/main/java/com/infinum/libname/libModule1/MyClass.java
@@ -1,0 +1,4 @@
+package com.infinum.libname.libModule1;
+
+public class MyClass {
+}
