+++ libModule2/src/main/java/com/infinum/libname/libModule2/MyClass.java
@@ -1,0 +1,4 @@
+package com.infinum.libname.libModule2;
+
+public class MyClass {
+}
