+++ sample/src/main/AndroidManifest.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.infinum.libname">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.infinum.libname.sample">
 
     <application android:allowBackup="true" android:icon="@mipmap/ic_launcher" android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round" android:supportsRtl="true" android:theme="@style/Theme.Androidlibrarytemplate">
-        <activity android:name="com.infinum.libname.MainActivity" android:exported="true">
+        <activity android:name="com.infinum.libname.sample.MainActivity" android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
